+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## vNext
+- `k8seventgenerationprocessor` Add support for parsing `VulnerabilityReport` manifests and generating `VulnerabilityDetail` entities and `VulnerabilityFinding` relationships.
 
 ## v0.140.2
 - No changes